LNP-287: ♻️  refactor drush cache:force-clear-pag & force-clear-all commands to be compatible with Drush 12.
