+++ docroot/modules/custom/page_cache_persist/page_cache_persist.services.yml
@@ -6,8 +6,3 @@ services:
   page_cache_persist.cache_router_rebuild_subscriber:
     class: Drupal\page_cache_persist\EventSubscriber\CacheRouterRebuildSubscriber
     decorates: cache_router_rebuild_subscriber
-  page_cache_persist.commands:
-    class: Drupal\page_cache_persist\Commands\PageCachePersistCommands
-    arguments: ['@cache.page']
-    tags:
-      - { name: drush.command }
